% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_data.R
\docType{data}
\name{cpt_data}
\alias{cpt_data}
\title{CPT_DATA}
\format{
An object of class \code{data.frame} with 100 rows and 16 columns.
}
\usage{
data(cpt_data)
}
\description{
A mock data set of a Cigarette Purchase Task (CPT), consisting of 15 items in total. The prices range from 'FREE ($0)'
to '$10' per cigarette. Specifically: c("0","0.05","0.10","0.20","0.30","0.40","0.50","0.75","1","2","3","4","5","7.5","10").
The items are administered to participants in arrays of 3. A zero consumption response within an array will end the instrument
after the final item within the array. Thus, this is an example of a partially-administered purchase task.
}
\examples{
\dontrun{
### --- The CPT data can be loaded into the R Environment using the `data()` function.

data(cpt_data)
}
}
\keyword{datasets}
