% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_curve.R
\name{elasticity_curve}
\alias{elasticity_curve}
\title{ELASTICITY CURVE}
\usage{
elasticity_curve(
  pt,
  id_var,
  k_span = c(2, 3, 4),
  eq_type = "koff",
  agg_type = "mean",
  id_curve = FALSE,
  id_label = TRUE
)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}

\item{k_span}{The k-values to test for curve fitting, with the default set to c(2,3,4)}

\item{eq_type}{Equation type to fit elasticity curve using the {beezdemand} package.
One of c("hs","koff"), with "hs" for Hursh and Silberbeg (2008) or Koffarnus,
Franck, Stein, and Bickel (2015).}

\item{agg_type}{Aggregate type to fit elasticity curve using the {beezdemand} package.
One of c("mean","pooled"), with the default set to "mean".}

\item{id_curve}{If set to TRUE, each individual curve is visualized on a single plot, with
those sensitive to price increases (i.e. show more elasticity) highlighted. Specifically,
highlighted individuals are those greater than 3 standard deviations away from the mean (z-score > 3).}

\item{id_label}{If set to TRUE, the ID labels of those identified as sensitive to price increases
are labelled on the plot, and their curves are highlighted. If FALSE, then the ID labels are provided
as a print out in the console. By default, id_label is set to TRUE.}
}
\description{
Models the elasticity curve, and calculates other values of interest such as
derived omax and pmax values, by fitting the purchase task data using the
{beezdemand} package. The overall sample curve is visualized, with the option
to visualize each individual curve on the same plot, identifying thoes with
extreme sensitivity to price (i.e. high elasticity values > a z-score of 3).
}
\examples{
\dontrun{
## Basic calculation of elasticity curve and derived values using default settings
pt2 <- elasticity_curve(pt, id_var = "ID")
}
}
