% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_elasticity.R
\name{pt_elasticity}
\alias{pt_elasticity}
\title{PT ELASTICITY}
\usage{
pt_elasticity(pt, id_var, type, k = NULL, y_type = "log10")
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}

\item{type}{The type of elasticity curve to derive from the data, one of c("overall","individual"). The default is "overall" which
will calculate overall elasticity for the entire data frame.}

\item{k}{The k-value to use for curve fitting. The default is NULL, in which the k-value is calculated from the entire sample, with
k representing the span of consumption in log10 units. Otherwise, a single numerical value can be given as the k-value, allowing for
comparisons across studies, since k influences the calculation of elasticity.}

\item{y_type}{The way the y-axis (prices) are to be displayed. One of c("log10","original"), with "log10" as the default, or
"original" for the existing price values.}
}
\value{
A ggplot2 graphical object; For `type` "individual", the original pt data frame plus the derived values for each individual
is returned.
}
\description{
This function derives and visualizes elasticity using the Koffarnus, Franck, Stein, and Bickel (2015) equation. Elasticity can be
derived for the entire sample or for individuals. This function will also visualize the demand curve(s).
}
