% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_empirical.R
\name{pt_empirical}
\alias{pt_empirical}
\title{`pt_empirical()`}
\usage{
pt_empirical(pt, id_var)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}
}
\value{
A data frame
}
\description{
This function extracts the empirical demand indices (Intensity, Breakpoint, Omax, and Pmax) from the purchase task.
}
\examples{

### --- Load Data
data("cpt_data")

### --- Prep Data
pt <- price_prep(cpt_data, id_var = "ID", vars = c(paste0("cpt",1:15)),
prices = c("0","0.05","0.10","0.20","0.30","0.40","0.50", "0.75","1","2","3","4","5","7.5","10"))

pt2 <- pt_prep(pt, id_var = "ID", remove0 = TRUE, max_val = 99)
pt3 <- pt_qc(pt2, id_var = "ID", type = "partial")

### --- Function Example
pt4 <- pt_empirical(pt3$data,id_var = "ID")

}
