% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_winsor.R
\name{pt_winsor}
\alias{pt_winsor}
\title{`pt_winsor()`}
\usage{
pt_winsor(
  pt,
  id_var,
  level = "price",
  z_val = 3.99,
  option = 3,
  index_var = NULL,
  delta = NULL
)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and relevant purchase task variables (prices or indicators).}

\item{id_var}{The name of the unique identifier (ID) as identified in the data frame.}

\item{level}{The level at which outlier management is occurring, one of c("price","indicator"). The default is "price", which will
winsorize all prices of the purchase task. The "indicator" option is for winsorization at the indicator level, in which the indicator
to be winsorized needs to be identified by the `index_var` argument.}

\item{z_val}{The absolute z-value to define outliers. The default is |3.99|.}

\item{option}{The winsorization option, one of c(1,2,3). The default outlier management technique is option 3.}

\item{index_var}{The name of the index variable to winsorize.}

\item{delta}{Used to retain winsorization order when using winsorization type 3 for level "indicator". At the price level, a
the 1 unit increase in consumption means that the delta value is by default 1. For winsorization at the "indicator" level, the delta
must be defined by the user. For elasticity, a small value of 0.001 is recommended.}
}
\value{
A list consisting of two data frames: "data" which consists of the `id_var` and `pt` including the winsorized value(s); and
"wins_table" which provides details on which value(s) by `id_var` were winsorized (values before and after provided).
}
\description{
This function helps users to manage outliers at the price level, or at the indicator level by using winsorization techniques. See details for the different options.
}
\details{
There are 3 winsorization options:

i) Option 1 replaces outliers with the maximum non-outlying value rounded up;
ii) Option 2 replaces outliers with a value 1 higher than highest (or 1 lower than the lowest) non-outlying value; or
iii) outlying values are replaced with 1 value above the next highest non-outlying value to maintain order.
}
\examples{

### --- Example Data
pt <- data.frame("ID" = c(1:36),
"Intensity" = c(10,12,15,0,99,11,7,6,12,7,8,10,5,6,10,0,3,
                7,5,0,2,3,5,6,10,15,12,7,0,9,0,6,7,8,4,5),
"Breakpoint" = c(1,2,5,0,10,3,0.5,0.2,0.3,3,4,5,7.5,0.5,2,0,0.1,
                 0.5,0.5,0,3,2,2,1,2,3,4,1,0,2,0,5,5,7.5,2,3))

### --- Function Example
pt2 <- pt_winsor(pt, id_var = "ID", level = "indicator", index_var = c("Intensity"), delta = 1)

}
