% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_corr.R
\name{pt_corr}
\alias{pt_corr}
\title{PT CORR}
\usage{
pt_corr(pt, heatmap = TRUE, alpha = 0.05, coef_size = 7)
}
\arguments{
\item{pt}{A data frame which includes only variables to be included in the correlation testing.}

\item{heatmap}{A logical argument on whether a heatmap should be produced. The default is TRUE.}

\item{alpha}{The alpha value to determine significance, in which significant p-values will have the corresponding correlation
coefficient in the heatmap in bold. The default alpha value is 0.05.}

\item{coef_size}{The text size of the coefficients in the heatmap, when the heatmap argument is set to TRUE. Default is 7.}
}
\value{
A data frame with correlation coefficients and p-values for each variable included in the correlation testing. If
heatmap is set to TRUE, then a ggplot2 graphical object is also returned.
}
\description{
This function will provide coefficients and p-values for a pairwise pearson correlation, and can also produce a heatmap
}
