% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_curve.R
\name{pt_curve}
\alias{pt_curve}
\title{PT CURVE}
\usage{
pt_curve(pt, id_var, type, k = NULL, y_type = "log10")
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}

\item{type}{The type of demand curve to derive from the data, one of c("overall","individual"). The default is "overall" which
will calculate overall demand for the entire data frame.}

\item{k}{The k-value to use for curve fitting. The default is NULL, in which the k-value is calculated from the entire sample, with
k representing the span of consumption in log10 units. Otherwise, a single numerical value can be given as the k-value,
allowing for comparisons across studies, since k influences the calculation of alpha. When k is calculated from the entire sample,
it is calculated using the mean average consumption at the lowest and highest price points and in instances when the mean average consumption
at the highest price point is 0, the lowest non-zero mean consumption is used.}

\item{y_type}{The way the y-axis (prices) are to be displayed. One of c("log10","original"), with "log10" as the default, or
"original" for the existing price values.}
}
\value{
A ggplot2 graphical object; For `type` "individual", the original pt data frame plus the derived values for each individual
is returned.
}
\description{
This function derives and visualizes alpha using the Koffarnus, Franck, Stein, and Bickel (2015) exponentiated equation. Alpha
(rate of change constant) can be derived for the entire sample or for individuals. This function will also visualize the demand curve(s).
}
