% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_empirical.R
\name{pt_empirical}
\alias{pt_empirical}
\title{PT EMPIRICAL}
\usage{
pt_empirical(pt, id_var)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}
}
\value{
A data frame
}
\description{
This function extracts the empirical demand indices (Intensity, Breakpoint, Omax, and Pmax) from the purchase task.
}
