% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_linear.R
\name{pt_linear}
\alias{pt_linear}
\title{`pt_linear()`}
\usage{
pt_linear(pt, id_var, type, zero_val = 0.001, group_var = NULL, co_var = NULL)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}

\item{type}{The level for fitting the demand curves, one of c("overall","group","individual"). The default is "overall" which will calculate
overall demand for the entire data frame. For `type` "group", equation-derived demand indicators for each of the groups as identified by the `group_var`
are visualized. When `type` equals "individual", equation-derived demand indicators are calculated for each individual as identified by `id_var`.}

\item{zero_val}{The value to substitute zero values in the price and consumption data (default is 0.001).}

\item{group_var}{The name of the grouping variable when `type` equals "group".}

\item{co_var}{Optional. A vector of covariates to add to the linear model.}
}
\value{
A ggplot2 graphical object; For `type` "individual", the original pt data frame plus the derived values for each individual is returned.
}
\description{
This function fits a linear demand curve in order to calculate eta. It also allows for the addition of covariates.
}
\examples{

##### Load Data
data("cpt_data")

##### Prep Data
pt <- price_prep(cpt_data, id_var = "ID", vars = c(paste0("cpt",1:15)),
prices = c("0","0.05","0.10","0.20","0.30","0.40","0.50", "0.75","1","2","3","4","5","7.5","10"))

pt2 <- pt_prep(pt, id_var = "ID", remove0 = TRUE, max_val = 99)
pt3 <- pt_qc(pt2, id_var = "ID", type = "partial", bounce_type = "p2p")

##### Function Example
pt4 <- pt_linear(pt3$data, id_var = "ID", type = "individual")

}
