% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_prep.R
\name{pt_prep}
\alias{pt_prep}
\title{PT PREP}
\usage{
pt_prep(pt, id_var, max_val = NULL)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier (ID) as found in the data frame.}

\item{max_val}{Optional identification of a maximum allowed response for any given price point.}
}
\description{
This function helps users prep their data by i) assigning zero values for remaining
items not administered; ii) identifying missing data by ID; and (optional)
iii) re-assigning the maximum value allowed at any price point.
}
\examples{
\dontrun{
pt <- pt_prep(pt, id_var = "ID", max_val = 99)
}
}
