% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_qc.R
\name{pt_qc}
\alias{pt_qc}
\title{PT QC}
\usage{
pt_qc(pt, id_var, bounce_val = 0.1, rev_n = 1.5, cons_n = 2)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier (ID) as identified in the data frame.}

\item{bounce_val}{Expressed as a proportion, the bounce value is used as a
threshold to identify excessive inconsistencies in responses. The default bounce
value is 0.1. IDs exceeding this bounce value are removed.}

\item{rev_n}{The number of acceptable reversals from zero. The default number is 1.5,
which removes IDs with 2 or more reversals from zero.}

\item{cons_n}{The number of consecutive zeroes to signify a reversal from zero using
the {beezdemand} package. The default number of consecutive zeroes is 2.}
}
\description{
This function helps users to conduct quality control on purchase task data by
using the 3-Criterion Algorithm by Stein et al. (2015) to remove non-systematic data:
Specifically, this function identifies and removes IDs with: i) trend violations;
ii) excessive bounce ratios; and iii) excessive reversals in responses.
}
\details{
Stein, J. S., Koffarnus, M. N., Snider, S. E., Quisenberry, A. J., & Bickel, W. K. (2015).
Identification and management of nonsystematic purchase task data: Toward best practice.
Experimental and clinical psychopharmacology, 23(5), 377.
}
\examples{
\dontrun{
## Using default values:
pt1 <- pt_qc(pt, id_var = "ID")
## Customizing the third criterion (number of reversals):
## This identifies and removes IDs with any reversals (with a reversal defined as 1 zero value).
pt2 <- pt_qc(pt, id_var = "ID", rev_n = 0, cons_n = 1)
}
}
