% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_summary.R
\name{pt_summary}
\alias{pt_summary}
\title{PT SUMMARY}
\usage{
pt_summary(pt, pt_vars, group = NULL)
}
\arguments{
\item{pt}{A data frame consisting of the purchase variables `pt_vars` to summarize.}

\item{pt_vars}{The purchase task variables to summarize.}

\item{group}{An optional grouping variable to summarize the purchase task variables by.}
}
\value{
An object of type data frame
}
\description{
This function provides basic summary descriptives of the variables. This function can also provide summary by a grouping variable.
}
\examples{
### ---------- Data Prep:
cpt_data$AB <- ifelse(grepl("A",cpt_data$ID),"A","B")

### --------- Descriptive Stats by Group
data_summ <- pt_summary(cpt_data, pt_vars = c("cpt1","cpt2","cpt3"), group = "AB")
}
