% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_summary.R
\name{pt_summary}
\alias{pt_summary}
\title{`pt_summary()`}
\usage{
pt_summary(pt, pt_vars, group_var = NULL)
}
\arguments{
\item{pt}{A data frame consisting of the purchase variables `pt_vars` to summarize.}

\item{pt_vars}{The purchase task variables to summarize.}

\item{group_var}{An optional grouping variable to summarize the purchase task variables by.}
}
\value{
An object of type data frame
}
\description{
This function provides basic summary descriptives of the variables. This function can also provide summary by a grouping variable.
}
\examples{

### --- Example Data
pt <- data.frame("ID" = c(1:36),
"Intensity" = c(10,12,15,0,99,11,7,6,12,7,8,10,5,6,10,0,3,
                7,5,0,2,3,5,6,10,15,12,7,0,9,0,6,7,8,4,5),
"Breakpoint" = c(1,2,5,0,10,3,0.5,0.2,0.3,3,4,5,7.5,0.5,2,0,0.1,
                 0.5,0.5,0,3,2,2,1,2,3,4,1,0,2,0,5,5,7.5,2,3))

### --- Function Example
pt_summary(pt, pt_vars = c("Intensity","Breakpoint"))

}
