% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor_index.R
\name{winsor_index}
\alias{winsor_index}
\title{WINSOR INDEX}
\usage{
winsor_index(pt, index_var, z_val = 3.99, type = "option3", delta = 0.001)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{index_var}{The name of the index variable to winsorize.}

\item{z_val}{The absolute z-value to define outliers.}

\item{type}{Type of winsorization, specifically one of: c("option1","option2","option3")}

\item{delta}{used to retain winsorization order when using winsorization type "option3"}
}
\description{
This function helps users to manage outliers in the purchase task variables through
winsorization techniques. The winsorization options are identical to the options
available in the `winsor_price()` function. Specifically, i) Option 1 replaces
outliers with the maximum non-outlying price rounded up; ii) Option 2 replaces
outliers with a price 1 higher than highest (or 1 lower than the lowest) non-outlying
value; or iii) outlying values are replaced with 1 value above the next highest
non-outlying value to maintain order.
}
\examples{
\dontrun{
pt2 <- winsor_index(pt, index_var = "Intensity", type = "option3")
}
}
